Worked on the integration of speech-to-text within the conversation mode of the companion.
Added a brand new file for handling conversations. I added a folder that contains utilties related to the conversation system.

I've kept most of the main file intact, because prehaps we want to keep some of the code that's there.

There's a specific async/threading issue that is ocurring after running the script more than once. It seems like a process is being kept open beyond the time we exit the program. But it's not really impacting the execution, since it reboots whatever process was there.

There needs to be a buffer of time before the mic can listen to what else we are saying. I think it would keep the conversation a bit more natural, or we could also just make the TTS stop speaking.